+++ src/main/java/fr/minuskube/inv/internal/PlayerInvTask.java
@@ -14,7 +14,7 @@
  *    limitations under the License.
  */
 
-package fr.minuskube.inv;
+package fr.minuskube.inv.internal;
 
 import fr.minuskube.inv.content.InventoryContents;
 import fr.minuskube.inv.content.InventoryProvider;
@@ -27,7 +27,7 @@ import org.jetbrains.annotations.NotNull;
  *
  * @since 1.0
  */
-final class PlayerInvTask extends BukkitRunnable {
+public final class PlayerInvTask extends BukkitRunnable {
 
     @NotNull
     private final Player player;
